requirements.txt
PyMarvel==0.1.0
tqdm==4.23.4